Added support for DROP INDEX

+++ lib/percona_ar/query_builder.rb
@@ -13,6 +13,9 @@ class PerconaAr::QueryBuilder
   def add(sql)
     if sql =~ /^ALTER TABLE `?([^ `]*)`? (.*)/i
       @tables[$1.to_s] << get_sql_for($2)
+    elsif sql =~ /DROP INDEX/i
+      drop_clause, table = sql.split(/ ON /i)
+      @tables[table.delete('`')] << get_sql_for(drop_clause)
     end
     self
   end
@@ -21,6 +24,7 @@ class PerconaAr::QueryBuilder
 
   def get_sql_for(cmd)
     return cmd unless cmd =~ /DROP/i && !(cmd =~ /COLUMN/i)
+    return cmd if cmd =~ /DROP INDEX/i
     cmd.gsub(/DROP/i, "DROP COLUMN")
   end
